requirement.txt
Python 3.11.3

# Web framework
Flask==3.0.3
Werkzeug==3.1.3
python-dotenv==1.1.1

# LLM (GGUF)
llama-cpp-python==0.3.16

# Guardrail & NLP
torch==2.5.1
transformers==4.43.4

# RAG & Vector Search
numpy==1.26.4
faiss-cpu==1.8.0.post1
scikit-learn==1.7.2
sentencepiece==0.2.0
tokenizers==0.19.1
protobuf==5.28.2

# Database & Auth
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11

# Email reset password
sendgrid==6.12.5
